[hotfix] fix 252 s3-redshift for single table

+++ seatunnel-connectors-v2/connector-s3-redshift/src/main/java/org/apache/seatunnel/connectors/seatunnel/redshift/commit/S3RedshiftSinkAggregatedCommitter.java
@@ -295,7 +295,9 @@ public class S3RedshiftSinkAggregatedCommitter extends FileSinkAggregatedCommitt
                 mergeS3FileToRedshiftWithExternalTable(tempFilePath, filepath);
             }
 
-            fileSystemUtils.deleteFile(filepath);
+            if (filepath != null) {
+                fileSystemUtils.deleteFile(filepath);
+            }
             log.info("delete file {} ", filepath);
         }
